requirements.txt
# Scriptum API v2.5 - Python Dependencies
# Web Framework
Flask==3.1.2
flask-cors==6.0.2

# Configuration
python-dotenv==1.2.1

# HTTP Requests
requests==2.31.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0

# Optional: Rate Limiting (uncomment if implementing)
# Flask-Limiter==3.8.0

# Optional: Production Server (uncomment for production)
# gunicorn==23.0.0
# gevent==24.2.1
